feat: estilizar página de detalhe de to-do

+++ projeto_pos/src/app/todo-detail/todo-detail.component.ts
@@ -4,11 +4,14 @@ import { TodoService } from "../todo.service";
 import { Todo } from "../todo.model";
 import { FormsModule } from "@angular/forms";
 import { CommonModule } from "@angular/common";
+import { MatDialog } from "@angular/material/dialog";
+import { ConfirmationDialogComponent } from "../confirmation-dialog/confirmation-dialog.component";
+import { MaterialModule } from "../material.module";
 
 @Component({
   selector: "app-todo-detail",
   standalone: true,
-  imports: [FormsModule, CommonModule],
+  imports: [FormsModule, CommonModule, MaterialModule],
   templateUrl: "./todo-detail.component.html",
   styleUrl: "./todo-detail.component.css",
 })
@@ -18,7 +21,8 @@ export class TodoDetailComponent implements OnInit {
   constructor(
     private route: ActivatedRoute,
     private todoService: TodoService,
-    private router: Router
+    private router: Router,
+    private dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -47,16 +51,22 @@ export class TodoDetailComponent implements OnInit {
   }
 
   deleteTodo(): void {
-    if (this.todo.id) {
-      this.todoService.deleteTodoById(this.todo.id.toString()).subscribe(
-        (response) => {
-          console.log("To-do deleted successfully");
-          this.router.navigate(["/"]);
-        },
-        (error) => {
-          console.log("Error deleting To-do", error);
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent);
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        if (this.todo.id) {
+          this.todoService.deleteTodoById(this.todo.id.toString()).subscribe(
+            (response) => {
+              console.log("To-do deleted successfully");
+              this.router.navigate(["/"]);
+            },
+            (error) => {
+              console.log("Error deleting To-do", error);
+            }
+          );
         }
-      );
-    }
+      }
+    });
   }
 }
